Add files via upload
This is a python game 
Where the user gets to input rock, paper and scissors or quit. If the user decides to quit then the game will end.
if the user continues to input rock, paper and scissors.
The computer picks a random option to paly and will tell the user f they have won or lost.
At the end of the game the user can quit and results of how many times the user and computer won will be displayed.